refactor: Dto 클래스 디렉토리 구조 수정
