+++ back/src/main/java/codesquad/issuetracker/issue/IssueController.java
@@ -1,6 +1,10 @@
 package codesquad.issuetracker.issue;
 
 import codesquad.issuetracker.config.LoginInterceptor;
+import codesquad.issuetracker.issue.dto.request.IssueContentUpdateDto;
+import codesquad.issuetracker.issue.dto.request.IssueFilterDto;
+import codesquad.issuetracker.issue.dto.request.IssueTitleUpdateDto;
+import codesquad.issuetracker.issue.dto.response.IssueShowDto;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
