+++ back/src/main/java/codesquad/issuetracker/issue/IssueRepositoryCustom.java
@@ -1,5 +1,7 @@
 package codesquad.issuetracker.issue;
 
+import codesquad.issuetracker.issue.dto.request.IssueFilterDto;
+
 import java.util.List;
 
 public interface IssueRepositoryCustom {
