+++ back/src/main/java/codesquad/issuetracker/issue/dto/request/IssueTitleUpdateDto.java
@@ -1,4 +1,4 @@
-package codesquad.issuetracker.comment;
+package codesquad.issuetracker.issue.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class CommentUpdateDto {
+public class IssueTitleUpdateDto {
 
-    private String content;
+    private String title;
 }
