+++ back/src/main/java/codesquad/issuetracker/milestone/dto/request/MilestoneSaveDto.java
@@ -1,7 +1,8 @@
-package codesquad.issuetracker.milestone;
+package codesquad.issuetracker.milestone.dto.request;
 
+import codesquad.issuetracker.milestone.Milestone;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDate;
@@ -9,17 +10,16 @@ import java.time.format.DateTimeFormatter;
 
 @Getter
 @Setter
-@NoArgsConstructor
-public class MilestoneUpdateDto {
+@AllArgsConstructor
+public class MilestoneSaveDto {
 
     private String name;
     private String description;
     private String dueDate;
 
-    public Milestone toEntity(Long id) {
+    public Milestone toEntity() {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern(Milestone.DATE_TIME_FORMAT);
         Milestone.MilestoneBuilder milestoneBuilder = Milestone.builder()
-                .id(id)
                 .name(name)
                 .description(description);
