+++ back/src/main/java/codesquad/issuetracker/milestone/dto/request/MilestoneUpdateDto.java
@@ -1,7 +1,8 @@
-package codesquad.issuetracker.milestone;
+package codesquad.issuetracker.milestone.dto.request;
 
-import lombok.AllArgsConstructor;
+import codesquad.issuetracker.milestone.Milestone;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.time.LocalDate;
@@ -9,16 +10,17 @@ import java.time.format.DateTimeFormatter;
 
 @Getter
 @Setter
-@AllArgsConstructor
-public class MilestoneSaveDto {
+@NoArgsConstructor
+public class MilestoneUpdateDto {
 
     private String name;
     private String description;
     private String dueDate;
 
-    public Milestone toEntity() {
+    public Milestone toEntity(Long id) {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern(Milestone.DATE_TIME_FORMAT);
         Milestone.MilestoneBuilder milestoneBuilder = Milestone.builder()
+                .id(id)
                 .name(name)
                 .description(description);
