+++ back/src/main/java/codesquad/issuetracker/milestone/dto/response/MilestoneShowDto.java
@@ -1,5 +1,6 @@
-package codesquad.issuetracker.milestone;
+package codesquad.issuetracker.milestone.dto.response;
 
+import codesquad.issuetracker.milestone.Milestone;
 import lombok.Getter;
 
 import java.time.LocalDate;
