+++ back/src/main/java/codesquad/issuetracker/user/dto/response/AssigneeShowDto.java
@@ -1,14 +1,15 @@
-package codesquad.issuetracker.user;
+package codesquad.issuetracker.user.dto.response;
 
+import codesquad.issuetracker.user.User;
 import lombok.Getter;
 
 @Getter
-public class UserShowDto {
+public class AssigneeShowDto {
 
     private String loginId;
     private String profileImage;
 
-    public UserShowDto(User user) {
+    public AssigneeShowDto(User user) {
         this.loginId = user.getLoginId();
         this.profileImage = user.getProfileImage();
     }
