+++ back/src/main/java/codesquad/issuetracker/user/dto/response/UserShowDto.java
@@ -1,14 +1,15 @@
-package codesquad.issuetracker.user;
+package codesquad.issuetracker.user.dto.response;
 
+import codesquad.issuetracker.user.User;
 import lombok.Getter;
 
 @Getter
-public class AssigneeShowDto {
+public class UserShowDto {
 
     private String loginId;
     private String profileImage;
 
-    public AssigneeShowDto(User user) {
+    public UserShowDto(User user) {
         this.loginId = user.getLoginId();
         this.profileImage = user.getProfileImage();
     }
